+++ docs/README.md
@@ -38,8 +38,8 @@
 
 ### 📍 애플리케이션 관리 기능
 
-- [ ] 애플리케이션 진행 상황 관리 기능 (시작, 게임중, 끝)
-- [ ] 게임 단계에 맞는 진행 기능 (다리 생성단계, 건너는단계, 게임종료/재시작 단계)
+- [x] 애플리케이션 진행 상황 관리 기능 (시작, 게임중, 끝)
+- [x] 게임 단계에 맞는 진행 기능 (다리 생성단계, 건너는단계, 게임종료/재시작 단계)
 </details>
 
 ---
