+++ src/lib/const/inputViewMessage.js
