+++ src/lib/const/outputViewMessage.js
