+++ src/lib/const/varType.js
