+++ src/lib/model/MovementStatus.js
