+++ src/lib/utils/getVarType.js
@@ -1,4 +1,4 @@
-const varTypeConst = require("../const/varType");
+const varTypeConst = require("../../lib/const/varType");
 function getVarType(param) {
   return Array.isArray(param) ? varTypeConst.array : typeof param;
 }
