--- src/utils/func/convertToNumber.js
@@ -1,1 +1,0 @@
-const convertToNumber = (target) => {};
